+++ Demo/Demo/AppDelegate.m
@@ -1,5 +1,5 @@
 #import "AppDelegate.h"
-#import <PopupBridge/POPPopupBridge.h>
+#import <PopupBridge/PopupBridge-Swift.h>
 
 @interface AppDelegate ()
 
@@ -9,7 +9,6 @@
 
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-    [POPPopupBridge setReturnURLScheme:@"com.braintreepayments.popupbridgeexample"];
     return YES;
 }