+++ Package.swift
@@ -1,10 +1,10 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.8
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 import PackageDescription
 
 let package = Package(
     name: "PopupBridge",
-    platforms: [.iOS(.v9)],
+    platforms: [.iOS(.v14)],
     products: [
         .library(
             name: "PopupBridge",
@@ -14,8 +14,7 @@ let package = Package(
     targets: [
         .target(
             name: "PopupBridge",
-            exclude: ["PopupBridge-Framework-Info.plist"],
-            publicHeadersPath: "Public"
+            exclude: ["PopupBridge-Framework-Info.plist"]
         ),
     ]
 )