+++ Sources/PopupBridge/Public/PopupBridge/PopupBridge.h
@@ -1,9 +1,0 @@
-#import <UIKit/UIKit.h>
-
-//! Project version number for PopupBridge.
-FOUNDATION_EXPORT double PopupBridgeVersionNumber;
-
-//! Project version string for PopupBridge.
-FOUNDATION_EXPORT const unsigned char PopupBridgeVersionString[];
-
-#import <PopupBridge/POPPopupBridge.h>