+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # PopupBridge iOS Release Notes
 
+## unreleased
+
+* Convert PopupBridge to Swift
+* Breaking Changes
+  * Require iOS 14+, Xcode 14.3+, and Swift 5.8+
+  * Remove deprecated `POPPopupBridge.open(url:sourceApplication:)` & `POPPopupBridge.open(url:options:)` methods
+  * Remove `POPPopupBridge.set(returnURLScheme:)`
+  * Add required `urlScheme` parameter to `POPPopupBridge.init()`
+
 ## 1.2.0 (2021-01-22)
 
 * Exclude arm64 simulator architectures via Podspec